requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.7
pytest==7.4.0
PyYAML==6.0.1
requests==2.31.0
urllib3==2.0.4
